+++ CHANGELOG.md
@@ -11,11 +11,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added support for variadic functions.
 - Added support for cancelling compilation.
   - To enable cancelling a compilation, supply a `CancellationToken` to your `CompilationJob` object. You can request that the compilation be cancelled by cancelling the token. For more information, see [Task Cancellation](https://learn.microsoft.com/en-us/dotnet/standard/parallel-programming/task-cancellation).
-- Language Server: Fixed an error that would crash the language server if a `YarnCommand`- and `YarnFunction`-tagged method was declared inside a nested C# class
+- Added `NodeDebugInfo.Range`, which contains the range in which a node appears in the source code.
+- Added `NodeDebugInfo.IsImplicit`, which indicates whether the node was created by the compiler (and does not appear in the source code).
 
 ### Changed
 
 - Improved the performance of the type checker when dealing with large numbers of variable declarations.
+- Fixed an issue where `Utility.TagLines` would add `#line:` tags to lines with `#shadow:` tags (which isn't allowed.)
+- Language Server: Fixed an error that would crash the language server if a `YarnCommand`- and `YarnFunction`-tagged method was declared inside a nested C# class
 
 ### Removed